Update protonyms_without_a_taxon_with_same_cleaned_name_only_subspecies.rb
It seems like most of the issues in this script could be done automatically to speed things up since there are a LOT of results. Not sure if Brian has mentioned this yet. I am going through the more complicated ones that can't be resolved automatically and fixing them. I am at the end of the 75 and need to get access to the rest of the list if I could please?